Fix bug in the RFC.
If, while inside a CCS, we "sniff" a CCS sent by another car and addressed to yet another car,
we must send a SCS so that they don't interfere with our blinking.

Note that the protocol *may* work even if we don't send that SCS: the other cars will start blinking
after at least X time, at which point we'll have already finished anyway. But the timing is probably
too strict to be able to rely on it.